u-boot-hostmobility-fw-utils: deploy fw_unlock_mmc.sh for tegra3 based machines
This is a utility script that will remove the "RO" mode on the mmc part
which contains the U-boot env. The script is put in profile.d which
means it will override the default fw_setenv command.

While at it removed a package post process for tegra2 that parsed
/proc/mtd to create the fw_config file. We intend to create these files
during build time.

Signed-off-by: Mirza Krak <mirza.krak@endian.se>