Create common .inc file for recipes-flexray-utils
Create common/revision.inc for the recipes that use the
hostmobility/mx-flexray-utils.git repo. Set a fixed revision to lock the
file versions to one that has been tested and verified. Note that the
source for flexware firmware is proprietary and the output
binary(mx-flexray.fw) is copied into the mx-flexray-utils repo.