MX-5: Add Co-CPU driver to kernel recipe
Modify Makefile and Kconfig files in kernel to include STM32 driver in
kernel. Driver is added to kernel zImage.

The method is to checkout the mxm5-cocpu git repo in an extra-drivers
folder at top level in kernel tree. We then create symbolic links in
drivers/platform to have the kernel Kbuild system pick up the sources.