MX5: Add Flash from USB functionality
Add a check from U-Boot to the CO-CPU that checks if the user has pushed
the reset button or otherwise set CO-CPU I2C register 2 to 1 to indicate that
the a re-programming from an image file shall be done. The check is done
in the bootcmd environmental variable.

If the MX5 shall be flashed, a script with the name flashmx5.scr(created
from flashmx5.cmd) is expected to be found on the first detected USB
flash memory. The script handles the flashing sequence by loading the
image file that shall be programmed on the eMMC of the MX5. The image
file shall have the name mx5-image.wic.gz

To trigger an update:
* Place flashmx5.cmd and mx5-image.wic.gz on an USB memory.
* Insert the USB memory into the MX-5
* Press the reset button
* Alternatively, use i2ctransfer -y 2 w3@0x46 0 2 1 and restart