Convert layer to kirkstone compatible
This layer will now only work for kirkstone or newer.
the major reason is that append/remove syntax and similar has been changed from '_' to ' :'

There is more to be done to make this layer working with our machines but
this commit starts the change to newer bsp versions.

Also this layer will only work for poky builds not Angstrom the last
compatabile layer is hardknott.