Add gpio-poweroff handler that pulls GPIO4_IO19
Add gpio-poweroff handler that pulls GPIO4_IO19 to completely turn off
the board. NOTE: another pm_power_off handler is already registered and
gpio-poweroff.c needs to be patched to overwrite it.

The power-off functions seems to be the handled by the psci driver that
turns off the CPU but we pull the gpio and turn off the power instead.

Reference:
https://www.thegoodpenguin.co.uk/blog/an-overview-of-psci/

A question is should we allow power-off when KL15/start-signal is on.