HMX:Put back systemd-networkd
Variscite wants to use NetworkManager, we want to use systemd-networkd.
Since the meta-variscite-bsp/recipes-core/systemd/systemd_%.bbappend
removes networkd, we cannot re-append it so we copy the bbappend recipe
to our layer and BBMASK the variscite version.

(Tried forcevariable PACKAGECONFIG but that does not seem to work with
kirstone, maybe another syntax?)