HMX: device tree: Add no-1-8-v to prevent sd card from runing on 1.8v
With the new rev 6 card we have set the power supply to NVCC_SD2_1V8_3V3
(LDO5) which use a vselect pin to set the sd card to 1.8v or
3.3V our issue is when it set 1.8v the pinctrl tuning does
not work for 1.8V and the result error is -110 (4 times).
Before rev 6 we used only 3.3V so this should be the same thing.
And if runing on a rev 5 this flag vill not do anything to it.