HMX/tcan4x5x: Add suspend/resume for Linux 6.1
Add a flag to the suspend class function that leaves the chip in a
running state with rx interrupt enabled, so that m_can device driver can
configure and use the interrupt as a wakeup source.

TODO:Set interface in sleep mode and restore registers on resume.

replaces:
meta-hostmobility-bsp/dynamic-layers/recipes-kernel/linux/linux-variscite/
0021-HMX-Add-suspend-to-backported-m_can-driver.patch
for Linux 6.1 kernel

Patch from:

https://www.spinics.net/lists/linux-can/msg15153.html
https://www.spinics.net/lists/linux-can/msg15154.html