hmx: add systemd poweroff changes to make the system shutdown correct
Change systemd-poweroff.service to halt instead of poweroff that makes the PMIC to not turn off. still the system unmount and sync like before. it is possible to test the system with 'halt' with or without driver command echo 5 > /sys/class/host_watchdog/ctl_shutdown.

Add host-watchdog-poweroff.service
It require the host-watchdog-driver( and  firmware) to safely shutdown the unit.
The command execute shutdown on m7(cocpu) after 5 seconds and if the system has start signal set (high),
the unit will instead of shutdown reboot. This behavior is the same for older Host mobility products.
Shutdown(cocpu) will release the internal start signal and after a timeout time it will toggle reset pin.
This service will only work with system 'halt' or a systemd-poweroff.service with halt instead of shutdown.