hm-autostart: Add a workaround on to early host_watchdog start message
Before this the hmx had no issue with it but the hmm has for some reason it starts the autostart to fast even that I tested to set it to later start. This check if the rpmsg device handler is up before launch the ctrl_start that disable the first watchdog timer and continue on heartbeat message with 5 sec timeout.