hmm: change systemd-poweroff to reboot the system
The base board does not have any way to restore itself when pmic is off. This is similar issue with the HMX but for the hmx it does have a way to shutdown (cut-off) with use of the m7(m4) co-cpu. This product variant right now does not have a start signal and therfore it can not enter shutdown.